% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations_stratified.R
\name{run_simulations_stratified}
\alias{run_simulations_stratified}
\title{Power calculation with Monte Carlo simulation}
\usage{
run_simulations_stratified(
  strata_sizes = list(very_preterm = 2500, moderate_preterm = 2500, late_preterm = 2500,
    not_preterm = 2500),
  n_simulations = 1000,
  n_parallel_cores = 5,
  p_preterm = list(very = 0.002, moderate = 0.05, late = 0.1),
  p_autism = 0.01,
  p_geriatric_preg = 1/8
)
}
\arguments{
\item{name}{description}
}
\value{
description
}
\description{
This function helps you calculate power for regressions
using Monte Carlo simulation with several variables
that potentially have interactions.
}
