% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Power calculation with Monte Carlo simulation}
\usage{
simulate(n_obs, p_preterm, p_autism, p_geriatric_preg)
}
\description{
Helper function for \code{run_simulations()}.
For more information see parameter definition there.
}
